+++ src/main/java/Model/SellingLog.java
@@ -3,23 +3,23 @@ package Model;
 import java.util.Date;
 
 public class SellingLog extends Log{
-    private Double totalPriceArrived;
-    private Double amountOfOff;
+    private int totalPriceArrived;
+    private int amountOfOff;
     private Product buyingProducts;
     private static int numberOfSellingLog;
 
-    public SellingLog(Date date, Double totalPriceArrived, Double amountOfOff, Product buyingProducts, Customer customer) {
+    public SellingLog(Date date, int totalPriceArrived, int amountOfOff, Product buyingProducts, Customer customer) {
         super("SellingLog" + numberOfSellingLog++, date, customer);
         this.totalPriceArrived = totalPriceArrived;
         this.amountOfOff = amountOfOff;
         this.buyingProducts = buyingProducts;
     }
 
-    public Double getTotalPriceArrived() {
+    public int getTotalPriceArrived() {
         return totalPriceArrived;
     }
 
-    public Double getAmountOfOff() {
+    public int getAmountOfOff() {
         return amountOfOff;
     }
